Added - AutoTool - AutoClutch - BlockESP skidded from skilled v3 - Fixed wTap - Added Hold and Toggle button to bind - I used the intellijs optimize code thing - made TargetHUD look a bit nicer

+++ src/main/java/keystrokesmod/client/clickgui/kv/components/KvCategoryComponent.java
@@ -33,7 +33,7 @@ public class KvCategoryComponent extends KvComponent {
 
     @Override
     public void draw(int mouseX, int mouseY) {
-    	y = (int) (actualY + KvModuleSection.categoryScroll);
+    	y = actualY + KvModuleSection.categoryScroll;
 		int color = KvModuleSection.moduleSec.currentCategory == this ? 0xFFFFFF00: isMouseOver(mouseX, mouseY) ? 0xA000FF00 : 0xFF00FF00;
 		FontUtil.normal.drawStringWithShadow(category.getName(), x, y, color);
