+++ src/main/java/me/PianoPenguin471/mixins/MixinEntity2.java
@@ -57,7 +57,7 @@ public abstract class MixinEntity2 {
        float f1 = MathHelper.sin(-yaw * 0.017453292F - (float)Math.PI);
        float f2 = -MathHelper.cos(-pitch * 0.017453292F);
        float f3 = MathHelper.sin(-pitch * 0.017453292F);
-       return new Vec3((double)(f1 * f2), (double)f3, (double)(f * f2));
+       return new Vec3(f1 * f2, f3, f * f2);
    }
 
 }
